requirements.txt
fastapi>=0.70.0
uvicorn>=0.15.0
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.0
pymysql>=1.0.2
python-dotenv>=0.19.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
pydantic>=1.9.0
pydantic_settings>=2.0.0
python-multipart>=0.0.5
alembic>=1.7.0
email-validator>=1.1.0
aiofiles>=0.7.0
requests>=2.26.0
msal>=1.16.0
pytz>=2021.3
jinja2>=3.0.0
python-dateutil>=2.8.0
schedule>=1.2.0 